Add some stub Dtos and reorganised models.
