+++ MangoRestaurant.Product.API/Models/Dtos/Base/BaseDto.cs
@@ -1,0 +1,7 @@
+﻿namespace MangoRestaurant.Product.API.Models.Dtos.Base
+{
+    public class BaseDto
+    {
+        public int Id { get; set; }
+    }
+}
