+++ MangoRestaurant.Product.API/Models/Dtos/ProductDto.cs
@@ -1,0 +1,8 @@
+﻿using MangoRestaurant.Product.API.Models.Dtos.Base;
+
+namespace MangoRestaurant.Product.API.Models.Dtos
+{
+    public class ProductDto : BaseDto
+    {
+    }
+}
