+++ MangoRestaurant.Product.API/Repositories/Interfaces/IProductRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace MangoRestaurant.Product.API.Repositories.Interfaces
+{
+    public interface IProductRepository
+    {
+    }
+}
